+++ src/Back-End/ProFin.Core/Models/PanelAlert.cs
@@ -4,11 +4,11 @@ public class PanelAlert
 {
     public string Budget { get; set; }
 
-    public double PercentageBudget { get; set; }
+    public decimal PercentageBudget { get; set; }
 
-    public double TotalValueUsed { get; set; }
+    public decimal TotalValueUsed { get; set; }
 
-    public double BudgetValue { get; set; }
+    public decimal BudgetValue { get; set; }
 
     public string Description { get; set; }
